Add turnline panel config entry & move user config util
